# lib/api-zod/src/generated/types/calcResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalcResultBreakdown } from "./calcResultBreakdown";
import type { Sport } from "./sport";

export interface CalcResult {
  calories: number;
  breakdown: CalcResultBreakdown;
  sport: Sport;
}