# lib/api-zod/src/generated/types/calcSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Sport } from "./sport";

export interface CalcSummary {
  totalSports: number;
  totalMultipliers: number;
  avgMet: number;
  topSportByMet?: Sport;
}