# lib/api-zod/src/generated/types/computeCaloriesBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ComputeCaloriesBody {
  sportId: number;
  durationMinutes: number;
  /** Fréquence cardiaque en bpm */
  heartRateBpm: number;
  /** VO2 max en ml/kg/min */
  vo2Max: number;
  /** Dénivelé positif en mètres */
  elevationGainMeters?: number;
  /** Poids en kg (optionnel, défaut 70) */
  weightKg?: number;
}