# lib/api-zod/src/generated/types/createSportBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSportBody {
  name: string;
  slug: string;
  baseMet: number;
  icon: string;
  appliesElevation: boolean;
}