# lib/api-zod/src/generated/types/multiplier.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Multiplier {
  id: number;
  /** Clé du paramètre (ex. duration, heart_rate, vo2_max, elevation) */
  paramKey: string;
  /** Nom affiché */
  label: string;
  /** Valeur du multiplicateur */
  value: number;
  description: string;
  minValue: number;
  maxValue: number;
  unit: string;
  updatedAt: Date;
}