# lib/api-zod/src/generated/types/sport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Sport {
  id: number;
  name: string;
  slug: string;
  /** Valeur MET de base du sport */
  baseMet: number;
  /** Emoji ou identifiant d'icône */
  icon: string;
  /** Si vrai, le dénivelé s'applique à ce sport */
  appliesElevation: boolean;
  createdAt: Date;
}